PR comments

+++ lib/view_models/users/overview.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:mml_admin/models/model_list.dart';
 import 'package:mml_admin/services/user.dart';
+import 'package:mml_admin/components/delete_dialog.dart';
+import 'package:mml_admin/components/progress_indicator.dart';
+import 'package:mml_admin/services/router.dart';
 
 /// View model for the users overview screen.
 class UsersOverviewViewModel extends ChangeNotifier {
@@ -18,14 +21,20 @@ class UsersOverviewViewModel extends ChangeNotifier {
 
   /// Deletes the users with the passed [userIds] or or aborts, if the user
   /// cancels the operation.
-  Future<bool> deleteUsers<int>(List<int> userIds) async {
-    try {
-      await _userService.deleteUsers(userIds);
-    } catch (e) {
-      // Do not reload list on error!
-      return false;
+  Future<bool> deleteUsers<int>(BuildContext context, List<int> userIds) async {
+    var shouldDelete = await showDeleteDialog(context);
+
+    if (shouldDelete) {
+      try {
+        showProgressIndicator();
+        await _userService.deleteUsers(userIds);
+        RouterService.getInstance().navigatorKey.currentState!.pop();
+      } catch (e) {
+        // Do not reload list on error!
+        return false;
+      }
     }
 
-    return true;
+    return shouldDelete;
   }
 }
